path specification in object files should be independent of DESTDIR
Wrapping Makefile.am in OpenHSP replaces the path specification in
`hsed.desktop` to put target files (e.g. hsed) where the original
project does not desire. This path should not include DESTDIR when
`make install`, otherwise, the target file path specifies a
relative path from where `install.bash` launches in a host.

In that reason, `--prefix` at `configure` must be an absolute path inside target
chroot.

TODO: we need the other measure to create or replace `hsed.desktop` to
enable `configure` to use relative `--prefix`.

1) add a parameter (a environment variable) to clarify chroot root.
autoconf accepts `--enable-feature[=arg]` or `--with-<package>[=arg]. The parameter can be a
candidate to distinguish if it is cross-build or not.

https://github.com/nghttp2/nghttp2/blob/00bd76fc3defeffc4fa5208598ef414ad5db56f1/android-config#L38

2) ask the OpenHSP project to place executables and data at `/usr/local/{bin,share}`.